Rework a little bit the package structure
