+++ src/main/scala/com/timeout/docless/schema/Required.scala
@@ -43,6 +43,18 @@ object Required {
         tFields.get
     }
 
+    implicit val genericCNil: Fields[CNil] = instance(Nil)
+
+    implicit def genericCoproduct[H, T <: Coproduct](
+      implicit hFields: Lazy[Fields[H]],
+      tFields: Fields[T]
+    ): Fields[H :+: T] = instance((hFields.value.get ++ tFields.get).distinct)
+
+    implicit def genericOutIsCooprod[A, R <: Coproduct](
+      implicit gen: Generic.Aux[A, R],
+      fields: Fields[R]
+    ): Fields[A] = instance(fields.get)
+
     implicit def genericFields[A, R](implicit gen: LabelledGeneric.Aux[A, R],
                                      rfs: Fields[R]): Fields[A] =
       instance(rfs.get)
